fip_create: Add --pad option
The `--pad N' option will make the size of the Firmware Image Package
a multiple of the given value by appending null bytes as needed.
This is needed on HiKey when a Device Tree Blob is included in the FIP,
because the recovery bootloader won't accept a file which size is not a
multiple of 4 bytes.
The default value is 1 (no padding).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>