Preliminary support for Hisilicon D02
The Hisilicon D02 development board has 16 Cortex-A57 cores in a SoC
called P660 or PV660 or hip05 or Hi1610ES. Although the board is
designed for 2 CPUs, only one is currently supported.

This commit contains changes in common code, which are tagged with
'#ifdef D02_HACKS' and need revisiting for better integration.

This was done as part of porting OP-TEE to D02.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>